[core] enable Interp_debug_TEST and flags with flag combinations
Now you can check for multiple flags, e.g.
if (Interp_debug_TEST(interp,
       PARROT_MEM_STAT_DEBUG_FLAG | PARROT_GC_DETAIL_DEBUG_FLAG))
Was used only for single flags, so no changes for existing code.
Document the single flag limitation for the left-over trace_flags test.
There's no -t flag combination, like verbose yet.

+++ src/gc/gc_private.h
@@ -48,8 +48,7 @@ extern int CONSERVATIVE_POINTER_CHASING;
 
 /*
 
-=item C<static void panic_failed_allocation(unsigned int line, unsigned long
-size)>
+=item C<static void panic_failed_allocation(unsigned int line, size_t size)>
 
 Print an error message and die.
